Improved representation of AmiraMesh and HyperSurface files
The structure of the header now better reflects the content of AmiraMesh
and Hyper surface files. The array declarations for example
	define Vertices 8
	define Nodes 12
	definne Lattice 3 5 7
are now liked to the descriptions of the corresponding data pointers for
example

	Vertices { float[3] Coordinates } @1

The latter is now accessible through the Coordinates attribute of the
Vertices Block of the AmiraHeader. The same applies to HyperSurface
files. The corresponding data streams are implicitly loaded on demmand
when the stream_data or the data attributes of the Coordinates block are
accessed for the first time. For the latter case the resulting
numpy.ndarray is implicitly reshaped such that it has as many rows as
defined by the array definition and as many columns as indicated by the
dimension of the data type for Coordinates elements. In the above
example the data attribute would return an array with 8 rows and 3
columns.

Further changes include that AmiraHeader is now itself a Block type
class. No dedicated DataStreams class is necessary any more.

A small side effect of all the above changes is that the module has been
ported to python3 while trying to keep compatibility with python2.