Simplified access to data, moved all properties to __getattributes__
Mixing attributes and properties can in python be a instable and tricky
beast. Especially when combined with a load on demmand scheme. This is
intensified by the fact that __getattr__ is only called when default
attribute lookup fails with an AttributeError exception. In case this
exception occurs while already executing property  getter method than
the property is reported to be unknown attribute and the true cause
of exception is lost.

This commit moves all property getter and __getattr__ methods to
__getattribute__ which is called unconditionally which has the
advantage that as long as no __getattr__ method is called all exceptions
thrown by code called from within __getattribute__ are left untouched
and thus stack trace is repored properly without any special handling.
Any new derived class therefore should also use __getattribute__ only
and ensure the latter is calling
            super(<classname>,self).__getattribute__(name)
in any case exempt when throwing an AttributeError.

Getter only properties in this context are mimicked by defining
__setattr__ method on root Block class, method which is called whenever
a value is assigned to any attribute. It prevents assignment from non
Block type class methods and functions.

A minor change introduced by this commit is that formatting now is
able to display backreferences to parent Block using python __repr__
style string only when directly printing child. To make this work
attribute has to be marked to be backreference by setting isparent
parameter of add_attr method to True.

        block.add_attr('name',parent_block,True)

If the third paremeter is set to False or ommited than parent_block
reference will be printed as if child of block.

Fixes small quoting bug in compiled regular expression used by
grammar.py