Fixed last inconsistencies and cleaned code
Previous commit had a bug in __setattr__ related to allow only
modification of attributes predeclared through __slots__. Only
the __slots__ of the class for which __setattr__ was called was
checked.

Python built in hasattr method relies on __getattribute__ to throw
AttributeError for detecting whether attribute exists or not.
On AmiraDataStream and HypeSurface blocks will allways trigger
loading and decoding of data especially when checking for
"data" and "stream_data" attributes. For HyperSurface files this
is not fully avoidable at least implicit loading of "stream_data"
but decoding of data could be avoided if not needed.
By implementing __contains__ allowing to check if specific attribute
with specified name is provided by block this can be handled properly.

Replaced Exceptions throw when an error was encountered in data file or
when end of file was reached by appropriate custom exceptions. This now
allows to report load on demmand attribute as not present if not found
in file and rethrow original AttributeError exception which triggered
load on demmand. While any other AttributeError and other Exception is
passed as is.

Added examples for on how to use new structure in tests/ and some mainly
binary example files in tests/data/ moved already existing example data
files to tests/data/