Refactor of dispatch time prefiltering for special ContentType files
While testing with old real data generated by Amira a reather unusually
formatted header caused problems properly loading the data. The header
of the 'tests/data/EqualLabeledColumnsSpreadSheet.am' file defines
9 arrays with exactly the same name 'AT' and it also defines nine data
streams with exactly the same name 'AT'. Only in the parameter section
it can be deduced that each of the 9 arrays represents one '__Column'
as the '__ColumnNameXXXX' parameters still read '__ColumnName0000'
through to '__ColumnName0009'. With all 9 arrays and corresponding data
haveing the same name only the last array was added to the AmiraHeader
structure. All others were droped.

This is solved by refactoring the dispatch time filtering of
array_definitions for collecting the array definitions which could be
part of a dedicated special 'ContentType' like a 'HxSpreadSheet'. The
filtering is now only applied if the special 'ContentType' parameter
is encountered while parsing the Parameters section of the file.
The main purpose of 'DispatchFilter' objects usde for filtering is two
fold.
  1) Indicate whether the content of the file needs a special handling
     during setup of the AmiraHeader File Structure
  2) Filter and rewrite the names of the arrays and the correponding
     data_names where necessary and update the affected low level
     array_defintion objects and data_defintion objects accordingly

The benefit of this refactoring is that the dispatch time filtering
became more stable and the interface could be further simplified.